[HACK] drm/rcar-du: Add default modes to VGA connector
According to the DRM developers drivers must not manually add default
modes to VGA connectors in an unknown connection state. Until a proper
solution is found (possibly by handling default modes in the DRM core,
as for connectors in the connected state), this hack is a convenience
that avoids forcing userspace to configure the modes manually.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>