net: virtio_net: enbale hdr size buffer in MRG_RXBUF negotiated
On rx and VIRTIO_NET_F_MRG_RXBUF is negociated, the buffer size
checking doesn't follow the virtio spec. The driver should accept a
buffer that size of is the hdr. Even if the buffer is smaller than
minimum packet size.

The virtio spec says:
If VIRTIO_NET_F_MRG_RXBUF is negotiated, each buffer MUST be at least
the size of the struct virtio_net_hdr.

Fixes: bcff3162f3e0 ("virtio_net: stricter short buffer length checks")
Signed-off-by: Shunsuke Mie <mie@igel.co.jp>