PCI: endpoint: Add new pci_epc_ops to align buffer address
Some pci ep controller has restriction to map pci memory on alignment. The
alignment is defference and dependeded controllers. In order to solve this
problem, this patch introduce a new pci_epc_ops->align_mem().

We can use the interface with buffer address and size to get aligned addres
and buffer size. The obtained buffer size is used to allocate epc memory.

Signed-off-by: Shunsuke Mie <mie@igel.co.jp>